requirements.txt
Flask
newspaper3k
spacy
nltk
scikit-learn
requests
python-dotenv
validators
gunicorn
sentence-transformers
simhash
responses
psutil
readability-lxml
beautifulsoup4
yake
redis
celery
SQLAlchemy
Flask-SQLAlchemy
psycopg2-binary
alembic
sentry-sdk[flask]
prometheus-client
structlog